requirements.txt
inflection==0.5.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
matplotlylib==0.1.0
numpy==1.25.2
pandas==2.0.3
Pillow==9.5.0
plotly==5.16.0
streamlit==1.26.0
streamlit-folium==0.13.0
folium==0.14.0
